Get Last N Dates

Generates a sequence of the last N dates in "YYYY-MM-DD" format.

Usage

get_last_n_dates(N = 30)

Arguments

N

An integer specifying the number of days to generate.

Value

A character vector of dates in "YYYY-MM-DD" format.

Examples

get_last_n_dates(30)
#>  [1] "2026-07-09" "2026-07-08" "2026-07-07" "2026-07-06" "2026-07-05"
#>  [6] "2026-07-04" "2026-07-03" "2026-07-02" "2026-07-01" "2026-06-30"
#> [11] "2026-06-29" "2026-06-28" "2026-06-27" "2026-06-26" "2026-06-25"
#> [16] "2026-06-24" "2026-06-23" "2026-06-22" "2026-06-21" "2026-06-20"
#> [21] "2026-06-19" "2026-06-18" "2026-06-17" "2026-06-16" "2026-06-15"
#> [26] "2026-06-14" "2026-06-13" "2026-06-12" "2026-06-11" "2026-06-10"